+++ BackEnd/src/main/java/org/sds/elevateconnect/mapper/IterationMapper.java
@@ -11,7 +11,7 @@ public interface IterationMapper {
     int insertIteration(Iteration iteration);
     int updateIteration(Iteration iteration);
     int deleteIteration(@Param("id") Integer id);
-    List<Iteration> getIterations(@Param("projectId") Integer projectId, Short status);
+    List<Iteration> getIterations(@Param("projectId") Integer projectId, Integer status);
     List<Iteration> getIterationsByProjectId(@Param("projectId") Integer projectId);
     Short getProjectStatusById(Integer id);
     int getIteratedTimeById(Integer id);
